Refactor server package to follow MVC structure and move config to internal
